Make fussy Windows compiler happy. Review URL: http://codereview.chromium.org/9178012
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@10416 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/spaces-inl.h
@@ -166,10 +166,10 @@ Page* Page::Initialize(Heap* heap,
   Page* page = reinterpret_cast<Page*>(chunk);
   ASSERT(chunk->size() <= kPageSize);
   ASSERT(chunk->owner() == owner);
-  intptr_t object_bytes = page->ObjectAreaEnd() - page->ObjectAreaStart();
+  int object_bytes =
+      static_cast<int>(page->ObjectAreaEnd() - page->ObjectAreaStart());
   owner->IncreaseCapacity(object_bytes);
-  owner->AddToFreeLists(page->ObjectAreaStart(),
-                        static_cast<int>(object_bytes));
+  owner->AddToFreeLists(page->ObjectAreaStart(), object_bytes);
 
   heap->incremental_marking()->SetOldSpacePageFlags(chunk);
